Divide item details into smaller parts & aggregate same page components on a folder
